jobwrapper - add env-CMSRunAnalysis.sh
The new script env-CMSRunAnalysis.sh script is intended
to be shared with WMCore team.

We sould make the effort of keeping in this script,
which can eventually be moved to WMCore's repo,
all the shared code among CRAB and WMCore jobwrapper
scripts (namely gWMS-CMSRunAnalysis.sh / CMSRunAnalysis.sh
and submit_py3.sh).

The script name and its functions names are just proposals
and can be changed if we can come up with better names.