jobwrapper - Do not clean cmsRun's env
When we launch cmsRun, we should imitate what WMCore is
doing: do not clean the env, to preserve all the env
variables set for us by the workern node / pilot
and by cmsset_default.sh

Just a (small ?) difference with respect to WMCore:
we do not remove the COMP's python from the pythonpath.
This can be dangerous and should be thoroughly tested.